Fix path updating in sonarr and radarr integration

+++ modules/file_operations.py
@@ -363,7 +363,8 @@ def move_file_to_output(logger, debug, input_file_path, output_folder, folder_st
 
     os.makedirs(os.path.dirname(output_path), exist_ok=True)
     log_debug(logger, f"Moving file '{input_file_path}' to '{output_path}'")
-    shutil.move(input_file_path, output_path)
+    if os.path.exists(input_file_path):
+        shutil.move(input_file_path, output_path)
 
     return {
         "output_folder": new_folders,
